requirements.txt
asgiref==3.2.9
Django==3.0.7
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2020.1
django-heroku
sqlparse==0.3.1
boto3
django-storages